LayoutTests/http/tests/inspector/filesystem/request-directory-content-expected.txt

Tests requestDirectoryContent command.

errorCode: 2
entries: (null)

errorCode: 8
entries: (null)

errorCode: 0
entries:
  0:
    url: filesystem:http://127.0.0.1:8000/temporary/hoge/piyo.html
    name: piyo.html
    isDirectory: false
    isTextFile: true
    mimeType: text/html
    resourceType: Document
  1:
    url: filesystem:http://127.0.0.1:8000/temporary/hoge/fuga
    name: fuga
    isDirectory: true

errorCode: 1
entries: (null)

errorCode: 2
entries: (null)

All tests have done.
